Update : Changing Most Skill.ID from ushort to int Update : Most Derived Constructor from Skill to handle int "ID" and remove ushort cast. Update : Packet Lib should now be the only place where ID are sent as "ushort" Update : Dictionary to handle disabled skill now reflect this change to "int" ID.
Nothing should change in actual behavior when ID were cast to ushort before even storing them in Object.
Relations between subspell ID and spell ID could get better with this.
If you have spell ID that collide once converted to ushort it could be a trouble for using them in Quickbar or GameEffect Icon.
It Should be no trouble for "Mob Only" Spells that are cast only on mobs (that include charm and pet) or that doesn't spawn an effect icon !

git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk/GameServer@3355 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ spells/Spell.cs
@@ -344,7 +344,7 @@ namespace DOL.GS
         }
 
         public Spell(DBSpell dbspell, int requiredLevel, bool minotaur)
-			: base(dbspell.Name, (ushort)dbspell.SpellID, (ushort)dbspell.Icon, requiredLevel)
+			: base(dbspell.Name, dbspell.SpellID, (ushort)dbspell.Icon, requiredLevel)
 		{
 			m_description = dbspell.Description;
 			m_target = dbspell.Target;
@@ -396,7 +396,7 @@ namespace DOL.GS
 		/// <param name="spell"></param>
 		/// <param name="spellType"></param>
 		public Spell(Spell spell, string spellType) :
-			base(spell.Name, (ushort)spell.ID, (ushort)spell.Icon, spell.Level)
+			base(spell.Name, spell.ID, (ushort)spell.Icon, spell.Level)
 		{
 			m_description = spell.Description;
 			m_target = spell.Target;
